Moved currency functionality to separate file
